refactor: separate concerns
